window is reset as navigator immediately rather
The earlier logic seemed unnecessarily convoluted. It was registering a listener
and only setting the window as a navigator after the progress was completed. This was probably
needed earlier since there used to be a separate progress dialog in lobo.

Resetting it immediately doesn't seem to have any ill effects on the functionality.
However it helps usability immensely, since the window and progress are immediately visible now.

for #65