move the job finished handler logic to renderer context
Earlier, the iframe load algorithm would set the job finished handler
on the content window.

However, the content window is recreated when the document changes. It
used to work because of a race condition.

Setting the job finished handler on the renderer context is a better approach
because the renderer context doesn't change for a given frame.