graceful shutdown of JS Scheduler Thread
Instead of Thread.interrupt() + Thread.stop(), we use the `windowClosing` flag
and check for it periodically, both during Javascript execution and while
waiting for more tasks to arrive in the queue (via the poll method).

This won't stop the Thread immediately if it is waiting for some Network request,
etc. For that, we would need a long timeout and interrupt() + stop() combination.

This commit also makes Thread shutdown during finalization of `Window` non-blocking.