removed body background color
Specifying the body background color causes a number of _unrelated_ test
failures in grinder. This is because the background is not always drawn
to include all children (especially positioned renderables, floats, etc).

Removing this definition enables 879 progressions, which will help us focus
on the more critical test cases.

However, once the backgrounds gets painted correctly, it might be worthwhile
to renable the body background and check the tests. Tracked in gh-181