[workaround] Round off the Screen DPI
For GH-185

The CSS21 tests use lengths such as `0.25in`. When the real screen DPI
is an odd number, the conversion to pixels results in test failures because
of round-off errors.

As a quick workaround, the DPI is being rounded off to a multiple of 16.
A better solution should be made. See GH-185 for future ideas.